# internal/gateway/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Waveyard API
  version: 1.0.0
  description: |
    HTTP API for the Waveyard beat marketplace.

    Use **Authorize** to provide an access token as `Bearer <token>` for protected
    operations. Refresh tokens are stored in an HTTP-only `refresh_token` cookie.
  contact:
    name: Waveyard Engineering
servers:
  - url: /
    description: Current API host
security: []
...
paths:
...
  /specs:
    get:
      tags: [Catalog]
      operationId: listSpecs
      summary: Search and filter beats
      parameters:
        - { name: category, in: query, schema: { type: string, enum: [beat, sample] } }
        - { name: genres, in: query, description: Comma-separated genres, schema: { type: string } }
        - { name: tags, in: query, description: Comma-separated tags, schema: { type: string } }
        - { name: moods, in: query, description: Comma-separated moods, schema: { type: string } }
        - { name: instruments, in: query, description: Comma-separated instruments, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: key, in: query, schema: { type: string } }
        - { name: min_bpm, in: query, schema: { type: integer, minimum: 0 } }
        - { name: max_bpm, in: query, schema: { type: integer, minimum: 0 } }
        - { name: min_duration, in: query, schema: { type: integer, minimum: 0 } }
        - { name: max_duration, in: query, schema: { type: integer, minimum: 0 } }
        - { name: min_price, in: query, schema: { type: number, minimum: 0 } }
        - { name: max_price, in: query, schema: { type: number, minimum: 0 } }
        - { name: sort, in: query, schema: { type: string } }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - { name: per_page, in: query, schema: { type: integer, minimum: 1 } }
        - $ref: "#/components/parameters/CloudflareCountry"
        - $ref: "#/components/parameters/CountryCode"
      responses:
        "200":
          description: Paginated specs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSpecs"
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Catalog]
      operationId: createSpecLegacy
      summary: Legacy spec creation endpoint
      description: Returns 410. Use the `/spec-uploads` workflow instead.
      deprecated: true
      security: *bearerSecurity
      responses:
        "410":
          description: This endpoint has been replaced by the upload pipeline
          content:
            text/plain:
              schema:
                type: string
                example: multipart spec uploads are no longer supported; use /spec-uploads
        "401": { $ref: "#/components/responses/Unauthorized" }
  /specs/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Catalog]
      operationId: getSpec
      summary: Get a beat by UUID, short code, or slug
      parameters:
        - $ref: "#/components/parameters/CloudflareCountry"
        - $ref: "#/components/parameters/CountryCode"
      responses:
        "200":
          description: Spec details
          headers:
            X-Cache:
              description: Whether the response was served from the catalog cache
              schema: { type: string, enum: [HIT, MISS] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Spec" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [Catalog]
      operationId: updateSpec
      summary: Update owned beat metadata and cover image
      security: *bearerSecurity
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [metadata]
              properties:
                metadata:
                  type: string
                  description: JSON-encoded spec metadata
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: Updated spec
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Spec" }
        <<: *standardErrors
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Catalog]
      operationId: deleteSpec
      summary: Delete an owned beat
      security: *bearerSecurity
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        <<: *standardErrors
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /users/{id}/specs:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Catalog]
      operationId: getUserSpecs
      summary: List a producer's published specs
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - { name: per_page, in: query, schema: { type: integer, minimum: 1 } }
        - $ref: "#/components/parameters/CloudflareCountry"
        - $ref: "#/components/parameters/CountryCode"
      responses:
        "200":
          description: Producer specs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedSpecs" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned by login or refresh.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HTTP-only refresh session cookie set by login.
  parameters:
    ID:
      name: id
      in: path
      required: true
      description: Resource UUID. Catalog lookup also accepts a short code or slug.
      schema: { type: string }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1 }
...
    CloudflareCountry:
      name: CF-IPCountry
      in: header
      description: Cloudflare two-letter country code. IN selects INR display prices; every other value selects USD.
      schema: { type: string, minLength: 2, maxLength: 2, example: IN }
    CountryCode:
      name: X-Country-Code
      in: header
      description: Fallback country code when CF-IPCountry is absent. IN selects INR display prices; every other value selects USD.
      schema: { type: string, minLength: 2, maxLength: 2, example: IN }
  responses:
...
    NoContent:
      description: Operation completed with no response body
    BadRequest:
      description: Invalid input
      content: &errorContent
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
        text/plain:
          schema: { $ref: "#/components/schemas/PlainTextError" }
    Unauthorized:
      description: Missing or invalid authentication
      content: *errorContent
...
    NotFound:
      description: Resource not found
      content: *errorContent
    Conflict:
      description: Request conflicts with current resource state
      content: *errorContent
...
    InternalError:
      description: Unexpected server error
      content: *errorContent
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: invalid request body }
        details: { type: string }
    PlainTextError:
      type: string
      description: Error body written by net/http's http.Error. Some handlers currently put JSON-looking text in this plain-text body.
      example: invalid request body
...
    Money:
      type: object
      required: [amount_minor, amount_major, currency]
      properties:
        amount_minor: { type: integer, format: int64, description: Amount in the currency's smallest unit }
        amount_major: { type: number, format: double, description: Decimal amount in the currency's major unit }
        currency: { type: string, example: INR }
...
    Genre:
      type: object
      required: [id, name, slug, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        created_at: { type: string, format: date-time }
    LicenseOption:
      type: object
      required: [id, spec_id, type, name, price, price_money, display_price_money, features, file_types, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        spec_id: { type: string, format: uuid }
        type: { type: string, enum: [Basic, Premium, Trackout, Unlimited] }
        name: { type: string }
        price: { type: number, format: double }
        price_money: { $ref: "#/components/schemas/Money" }
        display_price_money: { $ref: "#/components/schemas/Money" }
        features: { type: array, items: { type: string } }
        file_types: { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PublicAnalytics:
      type: object
      required: [play_count, favorite_count, total_download_count, is_favorited]
      properties:
        play_count: { type: integer }
        favorite_count: { type: integer }
        total_download_count: { type: integer }
        is_favorited: { type: boolean }
    Spec:
      type: object
      required: [id, producer_id, producer_name, title, category, type, bpm, key, description, image_url, preview_url, price, price_money, display_price_money, duration, free_mp3_enabled, created_at, updated_at, processing_status]
      properties:
        id: { type: string, format: uuid }
        producer_id: { type: string, format: uuid }
        producer_name: { type: string }
        title: { type: string }
        category: { type: string, enum: [beat, sample] }
        type: { type: string }
        bpm: { type: integer }
        key: { type: string }
        description: { type: string }
        image_url: { type: string, format: uri }
        preview_url: { type: string, format: uri }
        price: { type: number, format: double }
        price_money: { $ref: "#/components/schemas/Money" }
        display_price_money: { $ref: "#/components/schemas/Money" }
        duration: { type: integer, description: Seconds }
        free_mp3_enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        licenses: { type: array, items: { $ref: "#/components/schemas/LicenseOption" } }
        genres: { type: array, items: { $ref: "#/components/schemas/Genre" } }
        tags: { type: array, items: { type: string } }
        moods: { type: array, items: { type: string } }
        instruments: { type: array, items: { type: string } }
        waveform_peaks: { type: array, items: { type: integer, format: int64 } }
        analytics: { $ref: "#/components/schemas/PublicAnalytics" }
        short_code: { type: string }
        slug: { type: string }
        processing_status: { type: string, enum: [pending, processing, completed, failed] }
    Pagination:
      type: object
      required: [total, page, per_page, limit, offset, total_pages]
      properties:
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        total_pages: { type: integer }
    PaginatedSpecs:
      type: object
      required: [data, metadata]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Spec" } }
        metadata: { $ref: "#/components/schemas/Pagination" }
...